+++ src/common/global.ts
@@ -3,10 +3,12 @@ import * as keytarType from 'keytar';
 import * as vscode from 'vscode';
 import { IConnection } from './IConnection';
 import { Constants } from './constants';
+import { ResultsManager } from '../resultsview/resultsManager';
 
 export class Global {
   public static keytar: typeof keytarType = getCoreNodeModule('keytar');
   public static context: vscode.ExtensionContext = null;
+  public static ResultManager: ResultsManager = null;
 
   public static get Configuration(): vscode.WorkspaceConfiguration {
     return vscode.workspace.getConfiguration(Constants.ExtensionId);