+++ webpack.config.js
@@ -1,0 +1,24 @@
+const path = require('path');
+
+module.exports = {
+	entry: {
+		index: './src/resultsview/scripts/index.ts'
+	},
+	module: {
+		rules: [
+			{
+				test: /\.tsx?$/,
+				use: 'ts-loader',
+				exclude: /node_modules/
+			}
+		]
+	},
+	resolve: {
+		extensions: ['.tsx', '.ts', '.js']
+	},
+	devtool: 'inline-source-map',
+	output: {
+		filename: '[name].js',
+		path: path.resolve(__dirname, 'media')
+	}
+};